+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 ### Changed
 * Increased raycasting distance from 1,000 studs to 15,000 studs.
+* ScrollingFrames now use AutomaticCanvasSize.
 ### Fixed
 * Holding any of the Alt keys and click on a part should no longer convert the terrain.